QueryThreadTracesSummary方法unitTime建议定义为uint64_t ，可能会发生数据溢出
# message auto-generated for no-merge-commit merge:
merge dev330 into master

QueryThreadTracesSummary方法unitTime建议定义为uint64_t ，可能会发生数据溢出

Created-by: insightDeveloper

Match-id-4a2ff69055ba27ac764be407e2881a6de7e934f5

+++ server/src/modules/timeline/handler/UploadFileHandler.cpp
@@ -40,7 +40,7 @@ void UploadFileHandler::ParseTask(UploadFileRequest request)
             singleFileData.stringArray.size());
         return;
     }
-    int textSize = request.params.text.size();
+    size_t textSize = request.params.text.size();
     if (textSize > MAX_STR_SIZE) {
         ServerLog::Error("Failed to parser upload file ", fileId, ",current content size is ", textSize, "max size is ",
             MAX_STR_SIZE);
